garbage authenticates

+++ services/question.go
@@ -5,8 +5,8 @@ import (
 	"main/context"
 )
 
-func ListQuestions(surveyId int64) []Question {
-	rows, err := context.Ctx.Db.Query("select id, description, title, options from questions where survey_id = $1", surveyId)
+func ListQuestions(surveyId int64, userId int64) []Question {
+	rows, err := context.Ctx.Db.Query("select id, description, title, options from questions where survey_id = $1 and user_id = $2", surveyId, userId)
 	defer rows.Close()
 	if err != nil {
 		log.Print(err)
@@ -22,10 +22,10 @@ func ListQuestions(surveyId int64) []Question {
 	return questions
 }
 
-func GetQuestion(questionId int64) Question {
+func GetQuestion(questionId int64, userId int64) Question {
 	var question Question
 	question.Id = questionId
-	err := context.Ctx.Db.QueryRow("select description, title, options from questions where id = $1", questionId).Scan(&question.Description, &question.Title, &question.Options)
+	err := context.Ctx.Db.QueryRow("select description, title, options from questions where id = $1 and user_id = $2", questionId, userId).Scan(&question.Description, &question.Title, &question.Options)
 	if err != nil {
 		log.Print(err)
 	}
@@ -33,10 +33,10 @@ func GetQuestion(questionId int64) Question {
 	return question
 }
 
-func CreateQuestion(surveyId int64) Question {
+func CreateQuestion(surveyId int64, userId int64) Question {
 	var question Question
 	question.Title = "New Question"
-	row := context.Ctx.Db.QueryRow("insert into questions (survey_id, title) values ($1, $2) returning id", surveyId, question.Title)
+	row := context.Ctx.Db.QueryRow("insert into questions (survey_id, title, user_id) values ($1, $2, $3) returning id", surveyId, question.Title, userId)
 	err := row.Scan(&question.Id)
 	if err != nil {
 		log.Print(err)
@@ -46,17 +46,17 @@ func CreateQuestion(surveyId int64) Question {
 	return question
 }
 
-func UpdateQuestion(question Question) {
+func UpdateQuestion(question Question, userId int64) {
 	var surveyId int64
-	err := context.Ctx.Db.QueryRow("update questions set title = $1, description = $2, options = $3 where id = $4 returning survey_id", question.Title, question.Description, question.Options, question.Id).Scan(&surveyId)
+	err := context.Ctx.Db.QueryRow("update questions set title = $1, description = $2, options = $3 where id = $4 and user_id = $5 returning survey_id", question.Title, question.Description, question.Options, question.Id, userId).Scan(&surveyId)
 	if err != nil {
 		log.Println(err)
 	}
-	context.Ctx.Db.Exec("update surveys set updated = now() where id = $1", surveyId)
+	context.Ctx.Db.Exec("update surveys set updated = now() where id = $1 and user_id = $2", surveyId, userId)
 }
 
-func DeleteQuestion(questionId int64) {
+func DeleteQuestion(questionId int64, userId int64) {
 	var surveyId int64
-	context.Ctx.Db.QueryRow("delete from questions where id = $1 returning survey_id", questionId).Scan(&surveyId)
-	context.Ctx.Db.Exec("update surveys set updated = now(), questions_order = array_remove(questions_order, $1) where id = $2", questionId, surveyId)
+	context.Ctx.Db.QueryRow("delete from questions where id = $1 and user_id = $2 returning survey_id", questionId, userId).Scan(&surveyId)
+	context.Ctx.Db.Exec("update surveys set updated = now(), questions_order = array_remove(questions_order, $1) where id = $2 and user_id = $3", questionId, surveyId, userId)
 }
